Refactor in seperate files
