+++ apps/frontend/src/modules/TFChain/services/tfchainService.ts
@@ -181,95 +181,3 @@ export const doesFarmExistByName = async (name: string): Promise<boolean> => {
     const readableFarm = farm.toJSON();
     return readableFarm != 0;
 };
-
-// Method to check if a specific user by substrateId has their terms and conditions accepted
-// Returns a boolean
-export const hasAcceptedTermsAndConditions = async (id: string): Promise<boolean> => {
-    const api = await getSubstrateApi();
-
-    const listTermsAndConditions = (await api.query.tfgridModule.usersTermsAndConditions(id)).toJSON();
-
-    return Object.keys(listTermsAndConditions as Object).length > 0;
-};
-
-// Method to check if a specific user by substrateId has their terms and conditions accepted
-// Returns a boolean
-export const checkIfTermsAndConditionsAreAccepted = async (id: string, retries = 0): Promise<boolean> => {
-    const api = await getSubstrateApi();
-
-    while (retries < 5) {
-        const listTermsAndConditions = (await api.query.tfgridModule.usersTermsAndConditions(id)).toJSON();
-
-        if (Object.keys(listTermsAndConditions as Object).length > 0) {
-            return true;
-        }
-
-        await new Promise(resolve => setTimeout(resolve, 1000));
-        await checkIfTermsAndConditionsAreAccepted(id, (retries += 1));
-    }
-
-    return false;
-};
-
-// Sign terms and condition to blockchain
-// Returns a boolean for success
-export const signTermsAndConditions = async (keyRing: KeyringPair): Promise<boolean> => {
-    const termsAndConditionsUrl = <string>flagsmith.getValue('farm_terms_and_conditions_url');
-    const api = await getSubstrateApi();
-
-    try {
-        const submittableExtrinsic = api.tx.tfgridModule.userAcceptTc(termsAndConditionsUrl, 'NO_HASH');
-        await submitExtrensic(submittableExtrinsic, keyRing);
-        return true;
-    } catch {
-        console.error('Unable to sign terms and conditions');
-        return false;
-    }
-};
-
-// Accept terms and conditions for a specific user by substrateId
-export const acceptTermsAndConditions = async (keyRing: KeyringPair) => {
-    // Using activation service to make sure we have balance available
-    await activationServiceForSubstrate(keyRing.address);
-
-    // Check if balance is available => if no throw error
-    const isBalanceAvailable = isSubstrateBalanceAvailable(keyRing.address);
-    if (!isBalanceAvailable) throw new Error('Unable to load substrate balance');
-
-    // Sign terms and conditions to blockchain
-    const signed = await signTermsAndConditions(keyRing);
-    if (!signed) throw new Error('Unable to sign terms and conditions');
-
-    // Getting terms and conditions
-    const accepted = await checkIfTermsAndConditionsAreAccepted(keyRing.address);
-    if (!accepted) throw new Error("Can't fetch signed terms and conditions");
-};
-
-export const addTwin = async (keyRing: KeyringPair): Promise<number> => {
-    const api = await getSubstrateApi();
-
-    const submittableExtrinsic = api.tx.tfgridModule.createTwin('127.0.0.1');
-    await submitExtrensic(submittableExtrinsic, keyRing);
-
-    const twinId = await getTwinId(keyRing.address);
-    if (twinId == 0) return 0;
-
-    return twinId;
-};
-
-export const getTwinId = async (id: string, retries = 0): Promise<number> => {
-    const api = await getSubstrateApi();
-
-    while (retries < 5) {
-        const twinId = (await api.query.tfgridModule.twinIdByAccountID(id)).toJSON();
-
-        if (twinId != 0) {
-            return twinId as number;
-        }
-
-        await new Promise(resolve => setTimeout(resolve, 1000));
-        await getTwinId(id, (retries += 1));
-    }
-
-    return 0;
-};
